% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{DiceRoll}
\alias{DiceRoll}
\title{Roll dice of your choosing}
\usage{
DiceRoll(m)
}
\arguments{
\item{m}{Dice of your choosing of "D20" - 20 sided dice, "D12" - 12 sided dice, "D10_1" 10 sided dice of values 0-9, "D10_10" 10 sided dice of values 00 - 90 by multiples of 10, "D8" 8 sided dice, "D6" 6 sided dice, "D4" 4 sided dice of values {1 2 3}, {1 2 4}, {1 3 4}, or {2 3 4}}
}
\value{
DiceRoll - a value of a selected dice
}
\description{
Roll a dice of your choosing (Description)
}
\details{
Roll a dice of your choosing and get a randomly generated number as the result (Details)
}
\examples{
DiceRoll("D20")
}
\author{
Dustin Duffy
}
